+++ core/class/_Commands.json
@@ -2,7 +2,6 @@
 	 "refresh"             	:{"LogicalId":"refresh"             ,"Name":"Rafraichir"                         ,"Order":0  ,"Type":"action","SubType":"other"  ,"Unite":null   ,"Config_infoName":null                 ,"Config_value":null      ,"Config_minValue":null,"Config_maxValue":null,"Config_tempHL":null ,"setValue":null                  ,"Display_param_step":null   ,"IsHistorized":0,"IsVisible":1}
 	,"IsOnLine"            	:{"LogicalId":"IsOnLine"            ,"Name":"Online"                             ,"Order":1  ,"Type":"info"  ,"SubType":"binary" ,"Unite":null   ,"Config_infoName":null                 ,"Config_value":null      ,"Config_minValue":null,"Config_maxValue":null,"Config_tempHL":null ,"setValue":null                  ,"Display_param_step":null   ,"IsHistorized":0,"IsVisible":0}
 	,"cur_temp"            	:{"LogicalId":"cur_temp"            ,"Name":"Temperature"                        ,"Order":10 ,"Type":"info"  ,"SubType":"numeric","Unite":"°C"   ,"Config_infoName":null                 ,"Config_value":null      ,"Config_minValue":null,"Config_maxValue":null,"Config_tempHL":null ,"setValue":null                  ,"Display_param_step":null   ,"IsHistorized":0,"IsVisible":1}
-	,"cur_tempH"           	:{"LogicalId":"cur_temp"            ,"Name":"Temperature"                        ,"Order":10 ,"Type":"info"  ,"SubType":"numeric","Unite":"°C"   ,"Config_infoName":null                 ,"Config_value":null      ,"Config_minValue":null,"Config_maxValue":null,"Config_tempHL":true ,"setValue":null                  ,"Display_param_step":null   ,"IsHistorized":0,"IsVisible":1}
 	,"cur_humi"            	:{"LogicalId":"cur_humi"            ,"Name":"Taux Humidité"                      ,"Order":11 ,"Type":"info"  ,"SubType":"numeric","Unite":"%"    ,"Config_infoName":null                 ,"Config_value":null      ,"Config_minValue":null,"Config_maxValue":null,"Config_tempHL":null ,"setValue":null                  ,"Display_param_step":null   ,"IsHistorized":0,"IsVisible":1}
 	,"eco_temp"            	:{"LogicalId":"eco_temp"            ,"Name":"Temp. eco"                          ,"Order":12 ,"Type":"info"  ,"SubType":"numeric","Unite":"°C"   ,"Config_infoName":null                 ,"Config_value":null      ,"Config_minValue":null,"Config_maxValue":null,"Config_tempHL":null ,"setValue":null                  ,"Display_param_step":null   ,"IsHistorized":0,"IsVisible":1}
 	,"eco_temp_consigne"   	:{"LogicalId":"eco_temp_consigne"   ,"Name":"Consigne Temp. eco"                 ,"Order":13 ,"Type":"action","SubType":"slider" ,"Unite":null   ,"Config_infoName":"eco_temp"           ,"Config_value":"#slider#","Config_minValue":10  ,"Config_maxValue":20  ,"Config_tempHL":null ,"setValue":"eco_temp"            ,"Display_param_step":0.5  ,"IsHistorized":0,"IsVisible":0}
