+++ plugin_info/info.json
@@ -5,7 +5,7 @@
 	"licence" : "AGPL",
 	"author" : "l3flo/bodbod",
 	"require" : "3.1.7",
-	"category" : "devicecommunication",	
+	"category" : "wellness",	
 	"changelog" : "https://github.com/bodbod44/jeedom-heatzyV2/blob/master/docs/#language#/changelog.md",
 	"changelog_beta" : "https://github.com/bodbod44/jeedom-heatzyV2/blob/beta/docs/#language#/changelog.md",
 	"documentation" : "https://github.com/bodbod44/jeedom-heatzyV2/blob/master/docs/#language#/index.md",
